[standalone]: Decouple dashcard menu types from the SDK bundle
