+++ frontend/src/metabase/embedding-sdk/types/plugins.ts
@@ -1,8 +1,8 @@
 import type { ReactNode } from "react";
 
-import type { DashCardMenuItem } from "metabase/dashboard/components/DashCard/DashCardMenu/dashcard-menu";
 import type { ClickAction } from "metabase/visualizations/types";
 
+import type { DashCardMenuItem } from "./dashcard-menu-item";
 import type { MetabaseQuestion } from "./question";
 
 export type MetabaseDataPointObject = {
